+++ api_rest/src/main/java/example/org/books_system/DeliveryStatusController.java
@@ -1,17 +1,21 @@
 package example.org.books_system;
+
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
-
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 import java.util.List;
-import java.util.stream.Collectors;
 
-// We add import for DefaultMessageSourceResolvable
-import org.springframework.context.MessageSourceResolvable;
 @RestController
 @Slf4j
 public class DeliveryStatusController {
@@ -21,40 +25,51 @@ public class DeliveryStatusController {
 
     @GetMapping("/getDeliveryStatuses")
     public ResponseEntity<List<DeliveryStatus>> getDeliveryStatuses() {
-        List<DeliveryStatus> deliveryStatuses = deliveryStatusService.getDeliveryStatuses();
-        return new ResponseEntity<>(deliveryStatuses, HttpStatus.OK);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            List<DeliveryStatus> deliveryStatuses = deliveryStatusService.getDeliveryStatuses();
+            return new ResponseEntity<>(deliveryStatuses, HttpStatus.OK);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PutMapping("/addDeliveryStatus")
     public ResponseEntity<String> addDeliveryStatus(@Valid @RequestBody DeliveryStatus deliveryStatus) {
-        log.info("A request to add delivery status was received: {}", deliveryStatus);
-        deliveryStatusService.addDeliveryStatus(deliveryStatus);
-        log.info("Author successfully added: {}", deliveryStatus);
-        return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            deliveryStatusService.addDeliveryStatus(deliveryStatus);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PostMapping("/updateDeliveryStatus")
     public ResponseEntity<String> updateDeliveryStatus(@Valid @RequestBody DeliveryStatus deliveryStatus) {
-        log.info("A request to update the delivery status has been received: {}", deliveryStatus);
-        deliveryStatusService.updateDeliveryStatus(deliveryStatus);
-        log.info("\n" +
-                "Delivery status successfully updated: {}", deliveryStatus);
-        return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            deliveryStatusService.updateDeliveryStatus(deliveryStatus);
+            return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @DeleteMapping("/deleteDeliveryStatus/{id}")
     public ResponseEntity<String> deleteDeliveryStatus(@PathVariable Long id) {
-        log.info("A request to delete the delivery status has been received with ID: {}", id);
-        deliveryStatusService.deleteDeliveryStatus(id);
-        log.info("Delivery status with ID {} deleted successfully", id);
-        return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            deliveryStatusService.deleteDeliveryStatus(id);
+            return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<String> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        String errorMessage = ex.getBindingResult().getAllErrors().stream()
-                .map(MessageSourceResolvable::getDefaultMessage)
-                .collect(Collectors.joining(", "));
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Validation error: " + errorMessage);
+    // Helper method to check user permissions
+    private boolean hasAccess(String requiredRole) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication != null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals(requiredRole));
     }
 }
