+++ api_rest/src/main/java/example/org/books_system/GenreController.java
@@ -1,22 +1,20 @@
 package example.org.books_system;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
-import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
-
-
-import org.springframework.context.MessageSourceResolvable;
 
 @RestController
 @Slf4j
@@ -27,39 +25,51 @@ public class GenreController {
 
     @GetMapping("/getGenres")
     public ResponseEntity<List<Genre>> getGenres() {
-        List<Genre> genres = genreService.getGenres();
-        return new ResponseEntity<>(genres, HttpStatus.OK);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            List<Genre> genres = genreService.getGenres();
+            return new ResponseEntity<>(genres, HttpStatus.OK);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PutMapping("/addGenre")
     public ResponseEntity<String> addGenre(@Valid @RequestBody Genre genre) {
-        log.info("A request to add a species has been received: {}", genre);
-        genreService.addGenre(genre);
-        log.info("Genres successfully added: {}", genre);
-        return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            genreService.addGenre(genre);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PostMapping("/updateGenre")
     public ResponseEntity<String> updateGenre(@Valid @RequestBody Genre genre) {
-        log.info("A species update request has been received: {}", genre);
-        genreService.updateGenre(genre);
-        log.info("Species successfully updated: {}", genre);
-        return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            genreService.updateGenre(genre);
+            return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @DeleteMapping("/deleteGenre/{id}")
     public ResponseEntity<String> deleteGenre(@PathVariable Long id) {
-        log.info("A request has been received to remove the genres with id: {}", id);
-        genreService.deleteGenre(id);
-        log.info("Genre with ID {} has been deleted successfully", id);
-        return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            genreService.deleteGenre(id);
+            return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<String> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        String errorMessage = ex.getBindingResult().getAllErrors().stream()
-                .map(MessageSourceResolvable::getDefaultMessage)
-                .collect(Collectors.joining(", "));
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Validation error: " + errorMessage);
+    // WIP: - Helper method to check user permissions
+    private boolean hasAccess(String requiredRole) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication != null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals(requiredRole));
     }
 }
