+++ api_rest/src/main/java/example/org/books_system/RoleController.java
@@ -1,21 +1,16 @@
 package example.org.books_system;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import javax.validation.Valid;
-import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
-
-// We add import for DefaultMessageSourceResolvable
 import org.springframework.context.MessageSourceResolvable;
 
 @RestController
@@ -27,33 +22,52 @@ public class RoleController {
 
     @GetMapping("/getRoles")
     public ResponseEntity<List<Role>> getRoles() {
-        List<Role> roles = roleService.getRoles();
-        return new ResponseEntity<>(roles, HttpStatus.OK);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            List<Role> roles = roleService.getRoles();
+            return new ResponseEntity<>(roles, HttpStatus.OK);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PutMapping("/addRole")
     public ResponseEntity<String> addRole(@Valid @RequestBody Role role) {
-        log.info("A request to add a role was received: {}", role);
-        roleService.addRole(role);
-        log.info("\n" +
-                "Role successfully added\n: {}", role);
-        return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            log.info("A request to add a role was received: {}", role);
+            roleService.addRole(role);
+            log.info("Role successfully added: {}", role);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Created!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @PostMapping("/updateRole")
     public ResponseEntity<String> updateRole(@Valid @RequestBody Role role) {
-        log.info("A role update request was received: {}", role);
-        roleService.updateRole(role);
-        log.info("Role successfully updated: {}", role);
-        return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            log.info("A role update request was received: {}", role);
+            roleService.updateRole(role);
+            log.info("Role successfully updated: {}", role);
+            return ResponseEntity.status(HttpStatus.OK).body("Updated!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @DeleteMapping("/deleteRole/{id}")
     public ResponseEntity<String> deleteRole(@PathVariable Long id) {
-        log.info("A request to delete a role was received with ID: {}", id);
-        roleService.deleteRole(id);
-        log.info("Role with ID {} deleted successfully", id);
-        return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication!= null && authentication.getAuthorities().stream().anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"))) {
+            log.info("A request to delete a role was received with ID: {}", id);
+            roleService.deleteRole(id);
+            log.info("Role with ID {} deleted successfully", id);
+            return ResponseEntity.status(HttpStatus.OK).body("Deleted!");
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
